refactor: :art: adjust structure with Header Componenet out folder
