--- src/common/components/Header/header.css
@@ -1,5 +1,0 @@
-header {
-  background-color: gray;
-  display: flex;
-  justify-content: space-around;
-}
