+++ src/common/pages/Recipes.jsx
@@ -4,7 +4,7 @@ import { CATEG_DRINKS, CATEG_MEALS,
   DRINKS, fetchAPI, FETCH_CATEG_D, FETCH_CATEG_M, MEALS } from '../../services';
 import CategoryButton from '../components/CategoryButton';
 import RecipeCard from '../components/RecipeCard';
-import Header from '../components/Header/Header';
+import Header from '../components/Header';
 import Footer from '../components/Footer/Footer';
 import Loading from '../components/Loading';
